requirements.txt
camelot_py==0.11.0
reportlab==4.0.4
